2003-08-27 Jay Fenlason <fenlason@redhat.com>
	* bgpd/bgp_routemap.c: attr->med is type u_in32_t, should be
	  compared with UINT32_MAX
	* ospfd/ospfd.c: remove redundant assert
	* zebra/rtadv.c: add missing include for zebra/rib.h

+++ bgpd/bgp_routemap.c
@@ -879,17 +879,17 @@ route_set_metric (void *rule, struct prefix *prefix,
 
 	  if (strncmp (metric, "+", 1) == 0)
 	    {
-	      if (bgp_info->attr->med/2 + metric_val/2 > ULONG_MAX/2)
-		bgp_info->attr->med = ULONG_MAX-1;
+	      if (bgp_info->attr->med/2 + metric_val/2 > UINT32_MAX/2)
+	        bgp_info->attr->med = UINT32_MAX-1;
 	      else
-		bgp_info->attr->med += metric_val;
+	        bgp_info->attr->med += metric_val;
 	    }
 	  else if (strncmp (metric, "-", 1) == 0)
 	    {
-	      if (bgp_info->attr->med <= metric_val) 
-		bgp_info->attr->med = 0;
+	      if (bgp_info->attr->med <= metric_val)
+	        bgp_info->attr->med = 0;
 	      else
-		bgp_info->attr->med -= metric_val;
+	        bgp_info->attr->med -= metric_val;
 	    }
 	}
     }
